Changed and need to replace Account

+++ Java2/src/ru/geekbrains/finmanager/Account.java
@@ -38,7 +38,7 @@ public class Account {
 		return description;
 	}
 	
-	public double getBalance() {
+	public BigDecimal getBalance() {
 		return balance;
 	}
